Rename directories to v1beta1 naming
